api/pyproject.toml
[tool.poetry]
name = "ddos_rcp"
version = "0.0.1"
description = "RCP API Demo for DDoS 2022"
authors = ["Kyosuke Miyamura <ask@386.jp>"]

[tool.poetry.dependencies]
python = ">=3.9,<3.11"

pytest = "^6.2.5"
python-dotenv = "^0.19.2"

fastapi = "^0.70.0"
fastapi-utils = "^0.2.1"
gunicorn = "^20.1.0"
uvicorn = {extras = ["standard"], version = "^0.15.0"}

sqlmodel = "^0.0.6"
alembic = "^1.7.5"
requests = "^2.27.1"
psycopg2-binary = "^2.9.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
run_server = "run:main"
